+++ src/renderer/src/ui/Request/Response/index.tsx
@@ -245,7 +245,7 @@ export function Response({
             </div>
           </div>
 
-          <div className="flex min-h-0 flex-1 flex-col overflow-hidden">
+          <div className="flex min-h-0 flex-1 flex-col overflow-auto">
             <SegmentedTabPanel value="body">
               <CodeEditor
                 readOnly